Add para in between 'Duty to notify registrar' header and the table
Add in para in between the 'Duty to notify registrar' header and the table:

Duty to Notify Registrar of Change in Particulars and Circumstances

Pursuant to the … ... particulars and circumstances.

"The notification is to be submitted in **GoBusiness Licensing Portal** via an amendment application. You need to tick the checkbox (for each section that you wish to amend) in order to update your records with MinLaw. Refer to the table below for more details on how to submit the amendment application."